Added Score and a HUD
Added a score ns. This updates the score and draws it to the screen. It is
responsible for triggering the puck to reset, though the puck ns actually does
the reset.

Added a HUD to show the state of the app as it makes it simpler to see problems
in the code changes.