Updated score to check if opponent scores
When player scores and the puck is reset it toggles direction towards the
player.
Now when the puck goes past the player it will give the opponent a point.

Updated the player so the player's position is held in the state. This
is (maybe) needed to detect if the puck hits the player's paddle. We could do
this by just using the mouse position directly which would mean less code, but
is bad in the long run because it becomes dependent on the mouse position an so
would be harder to refactor if we switch to keyboard keys. This is important if
we want to make it a two player game.