Improved HUD
Added keypress to toggle HUD on and off.
Add granularity to HUD so we can turn off parts of it.
Now displaying coordinate position next to objects
Added cross hairs to show mouse pointer position.
The cross hairs were inspired by this p5 sketch -
https://editor.p5js.org/crispysmokedweb/sketches/0yL_E5Ryi from this evernote
page:
https://www.evernote.com/shard/s58/client/snv?noteGuid=04fe4e6c-0228-48e7-8f64-b278d572baf0&noteKey=a26dc8925e85d479&sn=https%3A%2F%2Fwww.evernote.com%2Fshard%2Fs58%2Fsh%2F04fe4e6c-0228-48e7-8f64-b278d572baf0%2Fa26dc8925e85d479&title=Explore%2Bgenerative%2Bart%2Bwith%2Bpaper%2Band%2Bp5.js

Updated the boss to use :pos to be consistent with the player and puck.