Started to implement hitting the puck
Core -
Included size of puck to the state, so it is no longer a set of magic
numbers.
Renamed a pick-random-palette, random-style as it is a clearer name.
Hud - now shows the dimentions of the screen cornerns.
Player - added function to detect a hit. This causes the puck to change
direction. More work needs to be done on this function as it's buggy - it can
hold on to the puck.  I've come across similar problems in the past. I will fix
in a future release.