Added on screen help
Added some information to the player for what they can do.
This is as a series of message boxes that will appear over the game.
Maybe could change this in a future iteration from help, to messages as High
Scores, Paused Game, Game over could be set here.

Changed the styles slightly to allow the help messages to use a different style
to the rest of the game.