Updated the hitting of the Puck
Fixed a bug with the player hitting the puck.
Previously it toggled the puck direction, which caused the puck to sometimes
stick to the paddle. I had expected this and the correction I made is for the
player to always repulse the puck away.

An enhancement would be to check if you are hitting it from the wrong side and
then it will push it into the players end.

There is a bug - if you hit it too quickly it will not register the hit. Another
enhancement will be to change the speed by how hard you hit it.

I added similar code to the opponent, so now a primitive game can be played.

Hud was changed to show what's been currently developed.

Boss.clj also (mod)'s the angle so it doesn't become too large.