renamed help ns to messages to be more generic
Renamed help to messages because I display more than just the help.
Help could refer to what to do when things go wrong or if you don't know how to
play. Therefore help is both ambiguous and having an implied meaning to
different players of the game. Messages is purposely ambiguous - it's a general
place for messages/information.