Simplified the naming of the update functions for each namespace.
Renamed the functions because it was a tautological error to be calling:
thing/update-thing
self/update-self

This is simpler. There's less code and maybe in the future it will point the way
towards a possible polymorphic refactoring. As the state is immutable this could
be rewritten to allow concurrent updates to every namespace. It's using
ClojureScript, which is restricted to one thread, but it may change in the future.