Simplified the draw function.
Simplified the draw function to apply a list of drawing functions for each part of
the game against the game's state.

Earlier versions of this was very repetitive as it repeated passing the state
across to each draw function. This new implementation does the same, but feels
more idiomatic as it's a applied over a sequence of functions.