Some more refactoring for simplicity
Been meaning to move the state out of the core for a while. It seemed most
flexible to have it in its own namespace rather than an edn file.

Moved the game objects into another namespace - universe - the point of that,
again is to make the core smaller, plus more refactoring to come will mean that
I can manage the drawing order of items in the same place.

At the moment the order drawing is done in a cond, but it's clumsy. I can ->
filter sort-by :y to get the correct order in the future and this will be more
flexible if more objects/game effects are added.