Minor refactoring of the state
Moved the universe functions out of the default state, to make it cleaner.
The init function can populate fields with functions.

The initiate namespace no longer requires quil. This is good because it's
concentrating on the state, rather than the drawing. quil is all about drawing.
Init is all about state.

Updated the game-world/centre function. I want to rely on the state to determine
the centre, rather than getting it from quil. One reason is it will make testing
easier, however I think that's rarely reason enough to refactor code. The main
reason is it's relying on reading state more, therefore we can work towards
cutting the dependency on quil, making it more portable.